+++ flintrock/flintrock.py
@@ -1,11 +1,13 @@
 import os
 import posixpath
 import errno
+import json
 import resource
 import sys
 import shutil
 import textwrap
-import traceback
+import urllib.parse
+import urllib.request
 import warnings
 
 # External modules
@@ -65,8 +67,11 @@ def option_requires(
         requires_any: list=[],
         scope: dict):
     """
-    Raise an exception if an option's requirements are not met. If conditional_value
-    is not None, then only check the requirements if option is set to that value.
+    Raise an exception if an option's requirements are not met.
+
+    The option's requirements are checked only if the option has a "truthy" value
+    (i.e. it's not a "falsy" value like '', None, or False), and if its value is
+    equal to conditional_value, if conditional_value is not None.
 
     requires_all: Every option in this list must be defined.
     requires_any: At least one option in this list must be defined.
@@ -75,15 +80,18 @@ def option_requires(
     corresponding variable names (e.g. --option-a becomes option_a) and looking them
     up in the provided scope.
     """
-    if (conditional_value is None or
-            scope[option_name_to_variable_name(option)] == conditional_value):
+    option_value = scope[option_name_to_variable_name(option)]
+
+    if option_value and \
+            (conditional_value is None or option_value == conditional_value):
         if requires_all:
             for required_option in requires_all:
                 required_name = option_name_to_variable_name(required_option)
-                if required_name not in scope or scope[required_name] is None:
+                if required_name not in scope or not scope[required_name]:
                     raise UsageError(
-                        'Error: Missing option "{missing_option}" is required by '
-                        '"{option}{space}{conditional_value}".'.format(
+                        "Error: Missing option \"{missing_option}\" is required by "
+                        "\"{option}{space}{conditional_value}\"."
+                        .format(
                             missing_option=required_option,
                             option=option,
                             space=' ' if conditional_value is not None else '',
@@ -95,8 +103,9 @@ def option_requires(
                     break
             else:
                 raise UsageError(
-                    'Error: "{option}{space}{conditional_value}" requires at least '
-                    'one of the following options to be set: {at_least}'.format(
+                    "Error: \"{option}{space}{conditional_value}\" requires at least "
+                    "one of the following options to be set: {at_least}"
+                    .format(
                         option=option,
                         space=' ' if conditional_value is not None else '',
                         conditional_value=conditional_value if conditional_value is not None else '',
@@ -122,9 +131,10 @@ def mutually_exclusive(*, options: list, scope: dict):
         bad_option1 = used_options.pop()
         bad_option2 = used_options.pop()
         raise UsageError(
-            'Error: "{option1}" and "{option2}" are mutually exclusive.\n'
-            '  {option1}: {value1}\n'
-            '  {option2}: {value2}'.format(
+            "Error: \"{option1}\" and \"{option2}\" are mutually exclusive.\n"
+            "  {option1}: {value1}\n"
+            "  {option2}: {value2}"
+            .format(
                 option1=variable_name_to_option_name(bad_option1),
                 value1=scope[bad_option1],
                 option2=variable_name_to_option_name(bad_option2),
@@ -173,11 +183,12 @@ def cli(cli_context, config, provider):
 @click.option('--spark-version',
               help="Spark release version to install.")
 @click.option('--spark-git-commit',
-              help="Git commit hash to build Spark from. "
-                   "--spark-version and --spark-git-commit are mutually exclusive.")
+              help="Git commit to build Spark from. "
+                   "Set to 'latest' to build Spark from the latest commit on the "
+                   "repository's default branch.")
 @click.option('--spark-git-repository',
               help="Git repository to clone Spark from.",
-              default='https://github.com/apache/spark.git',
+              default='https://github.com/apache/spark',
               show_default=True)
 @click.option('--assume-yes/--no-assume-yes', default=False)
 @click.option('--ec2-key-name')
@@ -192,7 +203,7 @@ def cli(cli_context, config, provider):
 @click.option('--ec2-ami')
 @click.option('--ec2-user')
 @click.option('--ec2-spot-price', type=float)
-@click.option('--ec2-vpc-id', default='')
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.option('--ec2-subnet-id', default='')
 @click.option('--ec2-instance-profile-name', default='')
 @click.option('--ec2-placement-group', default='')
@@ -259,6 +270,16 @@ def launch(
             '--ec2-ami',
             '--ec2-user'],
         scope=locals())
+    # The subnet is required for non-default VPCs because EC2 does not
+    # support user-defined default subnets.
+    # See: https://forums.aws.amazon.com/thread.jspa?messageID=707417
+    #      https://github.com/mitchellh/packer/issues/1935#issuecomment-111235752
+    option_requires(
+        option='--ec2-vpc-id',
+        requires_all=[
+            '--ec2-subnet-id'
+        ],
+        scope=locals())
 
     if install_hdfs:
         hdfs = HDFS(version=hdfs_version)
@@ -267,10 +288,15 @@ def launch(
         if spark_version:
             spark = Spark(version=spark_version)
         elif spark_git_commit:
-            print("Warning: Building Spark takes a long time. "
-                  "e.g. 15-20 minutes on an m3.xlarge instance on EC2.")
-            spark = Spark(git_commit=spark_git_commit,
-                          git_repository=spark_git_repository)
+            print(
+                "Warning: Building Spark takes a long time. "
+                "e.g. 15-20 minutes on an m3.xlarge instance on EC2.")
+            if spark_git_commit == 'latest':
+                spark_git_commit = get_latest_commit(spark_git_repository)
+                print("Building Spark at latest commit: {c}".format(c=spark_git_commit))
+            spark = Spark(
+                git_commit=spark_git_commit,
+                git_repository=spark_git_repository)
         services += [spark]
 
     if provider == 'ec2':
@@ -298,12 +324,37 @@ def launch(
         raise UnsupportedProviderError(provider)
 
 
+def get_latest_commit(github_repository: str):
+    """
+    Get the latest commit on the default branch of a repository hosted on GitHub.
+    """
+    parsed_url = urllib.parse.urlparse(github_repository)
+    repo_domain, repo_path = parsed_url.netloc, parsed_url.path.strip('/')
+
+    if repo_domain != 'github.com':
+        raise UsageError(
+            "Error: Getting the latest commit is only supported "
+            "for repositories hosted on GitHub. "
+            "Provided repository domain was: {d}".format(d=repo_domain))
+
+    url = "https://api.github.com/repos/{rp}/commits".format(rp=repo_path)
+    try:
+        with urllib.request.urlopen(url) as response:
+            result = json.loads(response.read().decode('utf-8'))
+            return result[0]['sha']
+    except Exception as e:
+        raise Exception(
+            "Could not get latest commit for repository: {r}"
+            .format(r=repo_path)) from e
+
+
 @cli.command()
 @click.argument('cluster-name')
 @click.option('--assume-yes/--no-assume-yes', default=False)
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.pass_context
-def destroy(cli_context, cluster_name, assume_yes, ec2_region):
+def destroy(cli_context, cluster_name, assume_yes, ec2_region, ec2_vpc_id):
     """
     Destroy a cluster.
     """
@@ -318,7 +369,8 @@ def destroy(cli_context, cluster_name, assume_yes, ec2_region):
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
     else:
         raise UnsupportedProviderError(provider)
 
@@ -335,13 +387,15 @@ def destroy(cli_context, cluster_name, assume_yes, ec2_region):
 @cli.command()
 @click.argument('cluster-name', required=False)
 @click.option('--master-hostname-only', is_flag=True, default=False)
-@click.option('--ec2-region')
+@click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.pass_context
 def describe(
         cli_context,
         cluster_name,
         master_hostname_only,
-        ec2_region):
+        ec2_region,
+        ec2_vpc_id):
     """
     Describe an existing cluster.
 
@@ -368,7 +422,8 @@ def describe(
         search_area = "in region {r}".format(r=ec2_region)
         clusters = ec2.get_clusters(
             cluster_names=cluster_names,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
     else:
         raise UnsupportedProviderError(provider)
 
@@ -398,13 +453,14 @@ def describe(
 @cli.command()
 @click.argument('cluster-name')
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 # TODO: Move identity-file to global, non-provider-specific option. (?)
 @click.option('--ec2-identity-file',
               type=click.Path(exists=True, dir_okay=False),
               help="Path to SSH .pem file for accessing nodes.")
 @click.option('--ec2-user')
 @click.pass_context
-def login(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
+def login(cli_context, cluster_name, ec2_region, ec2_vpc_id, ec2_identity_file, ec2_user):
     """
     Login to the master of an existing cluster.
     """
@@ -422,7 +478,8 @@ def login(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
         user = ec2_user
         identity_file = ec2_identity_file
     else:
@@ -436,13 +493,14 @@ def login(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
 @cli.command()
 @click.argument('cluster-name')
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 # TODO: Move identity-file to global, non-provider-specific option. (?)
 @click.option('--ec2-identity-file',
               type=click.Path(exists=True, dir_okay=False),
               help="Path to SSH .pem file for accessing nodes.")
 @click.option('--ec2-user')
 @click.pass_context
-def start(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
+def start(cli_context, cluster_name, ec2_region, ec2_vpc_id, ec2_identity_file, ec2_user):
     """
     Start an existing, stopped cluster.
     """
@@ -460,7 +518,8 @@ def start(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
         user = ec2_user
         identity_file = ec2_identity_file
     else:
@@ -474,9 +533,10 @@ def start(cli_context, cluster_name, ec2_region, ec2_identity_file, ec2_user):
 @cli.command()
 @click.argument('cluster-name')
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.option('--assume-yes/--no-assume-yes', default=False)
 @click.pass_context
-def stop(cli_context, cluster_name, ec2_region, assume_yes):
+def stop(cli_context, cluster_name, ec2_region, ec2_vpc_id, assume_yes):
     """
     Stop an existing, running cluster.
     """
@@ -491,7 +551,8 @@ def stop(cli_context, cluster_name, ec2_region, assume_yes):
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
     else:
         raise UnsupportedProviderError(provider)
 
@@ -513,6 +574,7 @@ def stop(cli_context, cluster_name, ec2_region, assume_yes):
 @click.argument('command', nargs=-1)
 @click.option('--master-only', help="Run on the master only.", is_flag=True)
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.option('--ec2-identity-file',
               type=click.Path(exists=True, dir_okay=False),
               help="Path to SSH .pem file for accessing nodes.")
@@ -524,6 +586,7 @@ def run_command(
         command,
         master_only,
         ec2_region,
+        ec2_vpc_id,
         ec2_identity_file,
         ec2_user):
     """
@@ -551,7 +614,8 @@ def run_command(
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
         user = ec2_user
         identity_file = ec2_identity_file
     else:
@@ -575,6 +639,7 @@ def run_command(
 @click.argument('remote_path', type=click.Path())
 @click.option('--master-only', help="Copy to the master only.", is_flag=True)
 @click.option('--ec2-region', default='us-east-1', show_default=True)
+@click.option('--ec2-vpc-id', default='', help="Leave empty for default VPC.")
 @click.option('--ec2-identity-file',
               type=click.Path(exists=True, dir_okay=False),
               help="Path to SSH .pem file for accessing nodes.")
@@ -588,6 +653,7 @@ def copy_file(
         remote_path,
         master_only,
         ec2_region,
+        ec2_vpc_id,
         ec2_identity_file,
         ec2_user,
         assume_yes):
@@ -622,7 +688,8 @@ def copy_file(
     if provider == 'ec2':
         cluster = ec2.get_cluster(
             cluster_name=cluster_name,
-            region=ec2_region)
+            region=ec2_region,
+            vpc_id=ec2_vpc_id)
         user = ec2_user
         identity_file = ec2_identity_file
     else:
@@ -792,7 +859,7 @@ def set_open_files_limit(desired_limit):
 
 def main() -> int:
     if flintrock_is_in_development_mode():
-        warnings.simplefilter(action='error', category=DeprecationWarning)
+        warnings.simplefilter(action='always', category=DeprecationWarning)
         # warnings.simplefilter(action='always', category=ResourceWarning)
 
     set_open_files_limit(4096)
@@ -808,10 +875,6 @@ def main() -> int:
     except UsageError as e:
         print(e, file=sys.stderr)
         return 2
-    except Exception as e:
-        if not isinstance(e, Error):
-            # This not one of our custom exceptions, so print
-            # a traceback to help the user debug.
-            traceback.print_tb(e.__traceback__, file=sys.stderr)
+    except Error as e:
         print(e, file=sys.stderr)
         return 1